Rename module: eventstore -> timetrackingstore
