+++ libimagentrytimetrack/src/lib.rs
@@ -34,7 +34,7 @@ extern crate libimagentrytag;
 mod constants;
 pub mod error;
 pub mod timetracking;
-pub mod eventstore;
+pub mod timetrackingstore;
 pub mod iter;
 pub mod result;
 pub mod tag;
